requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
openai-agents>=0.1.0
httpx>=0.27.0
